move stuff to commands package
